Move deep linking stuff out
